requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
Django==1.9.1
django-celery==3.1.17
djangorestframework==3.3.2
kombu==3.0.35
pytz==2016.4
wheel==0.24.0
psycopg2==2.6.1
django-rest-swagger==0.3.5
